feat: polished EAS Types and docs
