+++ src/identity/index.ts
@@ -4,4 +4,9 @@ export { Name } from './components/Name';
 export { useAvatar } from './hooks/useAvatar';
 export { useName } from './hooks/useName';
 export { getEASAttestations } from './getEASAttestations';
-export type { EASSchemaUid, EASAttestation, EASChainDefinition } from './types';
+export type {
+  EASSchemaUid,
+  EASAttestation,
+  EASChainDefinition,
+  GetEASAttestationsOptions,
+} from './types';
