+++ src/identity/types.ts
@@ -1,18 +1,12 @@
 import { Address } from 'viem';
 
 /**
- * Ethereum Attestation Service (EAS) Attester Address
- * The Ethereum address of the attester who created the attestation.
- */
-type EASAttesterAddress = Address;
-
-/**
  * Ethereum Attestation Service (EAS) Schema Uid
  * The schema identifier associated with the EAS attestation.
  *
  * Note: exported as public Type
  */
-export type EASSchemaUid = Address;
+export type EASSchemaUid = `0x${string}`;
 
 /**
  * Ethereum Attestation Service (EAS) Attestation
@@ -21,7 +15,7 @@ export type EASSchemaUid = Address;
  * Note: exported as public Type
  */
 export type EASAttestation = {
-  attester: EASAttesterAddress; // the attester who created the attestation.
+  attester: Address; // the attester who created the attestation.
   decodedDataJson: string; // The attestation data decoded to JSON.
   expirationTime: number; // The Unix timestamp when the attestation expires (0 for no expiration).
   id: string; // The unique identifier of the attestation.
@@ -42,4 +36,16 @@ export type EASChainDefinition = {
   easGraphqlAPI: string; // EAS GraphQL API endpoint
   id: number; // blockchain source id
   schemaUids: EASSchemaUid[]; // Array of EAS Schema UIDs
+};
+
+/**
+ * Attestation Options
+ *
+ * Note: exported as public Type
+ */
+export type GetEASAttestationsOptions = {
+  schemas?: EASSchemaUid[];
+  revoked?: boolean;
+  expirationTime?: number;
+  limit?: number;
 };
